+++ includes/admin/class-llms-sendwp.php
@@ -138,7 +138,10 @@ class LLMS_SendWP extends LLMS_Abstract_Email_Provider {
 
 		}
 
-		return '<button class="llms-button-outline" id="llms-sendwp-connect"><i class="fa fa-paper-plane-o" aria-hidden="true"></i> Connect SendWP</button>';
+		return sprintf(
+			'<button class="llms-button-outline" id="llms-sendwp-connect"><i class="fa fa-paper-plane-o" aria-hidden="true"></i> %s</button>',
+			esc_html__( 'Connect SendWP', 'lifterlms' )
+		);
 	}
 
 	/**